+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/member/GatheringMemberAppender.java
@@ -2,6 +2,7 @@ package com.triplem.momoim.core.domain.member;
 
 import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.core.domain.gathering.model.Gathering;
+import com.triplem.momoim.core.domain.member.infrastructure.GatheringMemberRepository;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
