+++ momoim-external-api/src/main/java/com/triplem/momoim/api/gathering/service/GatheringService.java
@@ -7,8 +7,8 @@ import com.triplem.momoim.core.domain.gathering.dto.GatheringSearchOption;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringCategory;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringSubCategory;
 import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
-import com.triplem.momoim.core.domain.member.GatheringMemberRepository;
 import com.triplem.momoim.core.domain.member.dto.GatheringMemberDetail;
+import com.triplem.momoim.core.domain.member.infrastructure.GatheringMemberRepository;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import java.util.ArrayList;
